replica/database.hh: drop readers/multishard.hh include
Said header is included for two things:
* read_ahead (bool_class)
* reader_lifecycle_policy (interface)

The former is solved by extracting read_ahead into
readers/multishar_fwd.hh. The latter is solved by extracting
streaming_reader_lifecycle_policy into its own header. It is only used
by database.cc and view.cc, which can now include the new header file
instead of relying on database.hh to supply it.

This allows dropping the readers/multishard.hh include from
replica/database.hh, which transitively pulled in the massive
locator/abstract_replication_strategy.hh.

(cherry picked from commit f7b5263079ce98574a2b4fe6a87a56902df171fe,
originally part of scylladb/scylladb#29009)
